+++ test/providers/google/googleLive.test.ts
@@ -1,5 +1,5 @@
 import WebSocket from 'ws';
-import { GoogleMMLiveProvider } from '../../../src/providers/google/googleMultimodalLive';
+import { GoogleMMLiveProvider } from '../../../src/providers/google/googleLive';
 
 jest.mock('ws');
